Fixed buffer len error, added close channel, refactored for readability
